:recycle: [#2066] add ReadOnlyMixin
