+++ src/openzaak/utils/hooks.py
@@ -22,6 +22,10 @@ def get_lib_doc_excludes():
     """
     Exclude ConvertNoneMixin docstring from api spec generation
     """
-    from .serializers import ConvertNoneMixin, SubSerializerMixin
+    from .serializers import ConvertNoneMixin, SubSerializerMixin, ReadOnlyMixin
 
-    return default_get_lib_doc_excludes() + [ConvertNoneMixin, SubSerializerMixin]
+    return default_get_lib_doc_excludes() + [
+        ConvertNoneMixin,
+        SubSerializerMixin,
+        ReadOnlyMixin,
+    ]
